refactor : 공개 여부 수정 API를 User가 아닌 LifeMap에 구현 및 관련 코드 변경 (#69)
