+++ application/api/src/main/kotlin/io/raemian/api/goal/GoalService.kt
@@ -2,22 +2,32 @@ package io.raemian.api.goal
 
 import io.raemian.api.goal.controller.request.CreateGoalRequest
 import io.raemian.api.goal.controller.response.CreateGoalResponse
+import io.raemian.api.goal.controller.response.GoalResponse
+import io.raemian.api.lifemap.LifeMap
 import io.raemian.api.sticker.StickerService
 import io.raemian.api.support.RaemianLocalDate
 import io.raemian.api.tag.TagService
-import io.raemian.api.user.UserService
 import io.raemian.storage.db.core.goal.Goal
 import io.raemian.storage.db.core.goal.GoalRepository
+import io.raemian.storage.db.core.lifemap.LifeMapRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class GoalService(
-    private val userService: UserService,
     private val stickerService: StickerService,
     private val tagService: TagService,
     private val goalRepository: GoalRepository,
+    private val lifeMapRepository: LifeMapRepository,
 ) {
+
+    @Transactional(readOnly = true)
+    fun getById(id: Long): GoalResponse {
+        val goal = goalRepository.getById(id)
+        validateLifeMapPublic(goal.lifeMap)
+        return GoalResponse(goal)
+    }
+
     @Transactional
     fun create(userId: Long, createGoalRequest: CreateGoalRequest): CreateGoalResponse {
         val (title, yearOfDeadline, monthOfDeadLine, stickerId, tagId, description) = createGoalRequest
@@ -25,9 +35,10 @@ class GoalService(
         val deadline = RaemianLocalDate.of(yearOfDeadline, monthOfDeadLine)
         val sticker = stickerService.getById(stickerId)
         val tag = tagService.getById(tagId)
-        val user = userService.getById(userId)
+        val lifeMap = lifeMapRepository.findAllByUserId(userId)
+            .first()
 
-        val goal = Goal(user, title, deadline, sticker, tag, description!!, emptyList())
+        val goal = Goal(lifeMap, title, deadline, sticker, tag, description!!, emptyList())
         goalRepository.save(goal)
         return CreateGoalResponse(goal)
     }
@@ -39,8 +50,14 @@ class GoalService(
         goalRepository.delete(goal)
     }
 
+    private fun validateLifeMapPublic(lifeMap: LifeMap) {
+        if (!lifeMap.isPublic) {
+            throw SecurityException()
+        }
+    }
+
     private fun validateGoalIsUsers(userId: Long, goal: Goal) {
-        if (userId != goal.user.id) {
+        if (userId != goal.lifeMap.user.id) {
             throw SecurityException()
         }
     }
