+++ application/api/src/main/kotlin/io/raemian/api/lifemap/UpdatePublicationRequest.kt
@@ -1,0 +1,5 @@
+package io.raemian.api.lifemap
+
+data class UpdatePublicationRequest(
+    val isPublic: Boolean,
+)
