--- application/api/src/main/kotlin/io/raemian/api/user/controller/UpdateGoalsPublicRequest.kt
@@ -1,5 +1,0 @@
-package io.raemian.api.user.controller
-
-data class UpdateGoalsPublicRequest(
-    val isGoalsPublic: Boolean,
-)
